Active Android

+++ app/src/main/java/com/codepath/apps/twitterclient/TwitterClient.java
@@ -44,7 +44,7 @@ public class TwitterClient extends OAuthBaseClient {
 	 *    i.e client.post(apiUrl, params, handler);
 	 */
 
-    public void getHomeTimeline(Double maxId, AsyncHttpResponseHandler handler) {
+    public void getHomeTimeline(Long maxId, AsyncHttpResponseHandler handler) {
         String apiUrl = getApiUrl("statuses/home_timeline.json");
         RequestParams params = new RequestParams();
         params.put("count", 25);
